refactor components by appbar
